+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## 1.8.4.5 - 2024-02-29
+
+- Added `elao/enum` dependency.
+- Use `EnumCase` attribute for `PdfPageSize` class.
+- Modified the `PdfDocument` class:
+  - Removed the `getXY()` function.
+  - Added the `getPosition()` and `setPosition()` functions. 
+- Updated license.
+
 ## 1.8.4.4 - 2024-02-27
 
 - Updated license.